+++ src/components/swapv2/TokenInfo/utils.ts
@@ -2,6 +2,9 @@ import { t } from '@lingui/macro'
 import { SecurityInfo } from 'services/coingecko'
 
 import { ItemData, WarningType } from 'components/swapv2/TokenInfo/SecurityInfo/Content'
+import { TokenInfo } from 'hooks/useTokenInfo'
+import { formattedNum } from 'utils'
+import { formatLongNumber } from 'utils/formatBalance'
 
 export const RISKY_THRESHOLD = {
   RISKY: 0.05,
@@ -80,7 +83,7 @@ export const getSecurityTokenInfo = (data: SecurityInfo | undefined) => {
       type: WarningType.WARNING,
       riskyReverse: true,
     },
-  ]
+  ].filter(el => el.value !== undefined)
 
   const tradingData: ItemData[] = [
     {
@@ -141,10 +144,46 @@ export const getSecurityTokenInfo = (data: SecurityInfo | undefined) => {
       type: WarningType.WARNING,
       riskyReverse: true,
     },
-  ]
+  ].filter(el => el.value !== undefined)
 
   const { totalRisk: totalRiskContract, totalWarning: totalWarningContract } = calcTotalRisk(contractData)
   const { totalRisk: totalRiskTrading, totalWarning: totalWarningTrading } = calcTotalRisk(tradingData)
 
   return { contractData, tradingData, totalRiskContract, totalWarningContract, totalWarningTrading, totalRiskTrading }
+}
+
+const NOT_AVAILABLE = '--'
+export const getMarketTokenInfo = (tokenInfo: TokenInfo) => {
+  const listData = [
+    { label: t`Price`, value: tokenInfo.price ? formattedNum(tokenInfo.price.toString(), true) : NOT_AVAILABLE },
+    {
+      label: t`Market Cap Rank`,
+      value: tokenInfo.marketCapRank ? `#${formattedNum(tokenInfo.marketCapRank.toString())}` : NOT_AVAILABLE,
+    },
+    {
+      label: t`Trading Volume (24H)`,
+      value: tokenInfo.tradingVolume ? formatLongNumber(tokenInfo.tradingVolume.toString(), true) : NOT_AVAILABLE,
+    },
+    {
+      label: t`Market Cap`,
+      value: tokenInfo.marketCap ? formatLongNumber(tokenInfo.marketCap.toString(), true) : NOT_AVAILABLE,
+    },
+    {
+      label: t`All-Time High`,
+      value: tokenInfo.allTimeHigh ? formattedNum(tokenInfo.allTimeHigh.toString(), true) : NOT_AVAILABLE,
+    },
+    {
+      label: t`All-Time Low`,
+      value: tokenInfo.allTimeLow ? formattedNum(tokenInfo.allTimeLow.toString(), true) : NOT_AVAILABLE,
+    },
+    {
+      label: t`Circulating Supply`,
+      value: tokenInfo.circulatingSupply ? formatLongNumber(tokenInfo.circulatingSupply.toString()) : NOT_AVAILABLE,
+    },
+    {
+      label: t`Total Supply`,
+      value: tokenInfo.totalSupply ? formatLongNumber(tokenInfo.totalSupply.toString()) : NOT_AVAILABLE,
+    },
+  ]
+  return listData
 }
