+++ src/services/coingecko.ts
@@ -48,7 +48,7 @@ const coingeckoApi = createApi({
 })
 
 // todo tracking
-// todo (not for now) move basic chart api to this file
+// todo danh (not for now) move basic chart api to this file
 export const { useGetMarketTokenInfoQuery, useGetSecurityTokenInfoQuery } = coingeckoApi
 
 export default coingeckoApi
